disable flip logic in tile grid
