+++ packages/frontend/src/components/Game/GameBoard/Tile/index.jsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import './style.scss';
-import useGameContext from 'hooks/useGameContext';
 
 const Tile = ({ tile, onFlip }) => {
   const onClickTile = () => {
+    // if(state.f)
     onFlip(tile);
   };
-
-  const { state } = useGameContext();
 
   const tileClasses = [
     tile.faceUp ? 'faceUp' : 'faceDown tile border border-primary',
@@ -18,7 +16,7 @@ const Tile = ({ tile, onFlip }) => {
   return (
     <div
       className={tileClasses}
-      onClick={tile.faceUp || state.flipped.length > 1 ? null : onClickTile}
+      onClick={onClickTile}
       data-testid={`tile-${tile.id}`}
     >
       {!tile.faceUp ? null : (
